project: Disable backing up app data
The backing up feature allows transfering app settings between devices
(e.g. when you get a new phone).

Our app does not have static files that are stored locally and can be
transfered that way (everything is in the cloud), so keeping this does
not make much sense for now.

This fixes the AllowBackup: AllowBackup/FullBackupContent lint warning.