test: Remove unnecessary double tap when modifying a task
This test had 2 issues:

1. It was issuing a long click instead of a normal one, which has been
deprecated since we removed the need for long taps here.
2. That long/short click was nevertheless unnecessary, because the
actionOnItemAtPosition call already handles focusing the item.

Because of the second tap (see 2), the cursor was being moved in small
screens, causing the typed text to be placed in unwanted places. By
issuing a single tap, we make sure that the cursor is placed at the end
of the text (making tests deterministic across devices).